Update sitter filter form fields and typography
The changes introduce new fields to the sitter filter form (replacing "login" and "password") with "category", "date", and "address". We are also updating the typography variant in the content card from "body1" to "h5". With the changes to the sitter filter form, a new 'DateRange' type was introduced and the token is being removed when an error 401 occurs.