Update date field functionalities in sitter filter form
The date field validation has been disabled for improvement considerations. A new 'DateRangePickerFiled' component has been added to replace the existing TextField in the sitter filter form. This enhances date selection functionality. Also, the date field data in the 'onSubmit' function of the 'SitterFilter' is logged to the console. The 'DateRange' type is also updated to use 'luxon' DateTime type.