Add new Sitter page and update SitterCard component
A new page, Sitter, has been added which can be navigated to using the '/sitter/:sitterId' route. The SitterCard component has been updated to include a link to the specific sitter's page using the sitter ID. Changes also included minor styling adjustments and the addition of a new Link component for improved navigation.