Refactor sitter form data processing
A new module `data-to-variables` has been added in `src/entities/sitter/utils` to handle the conversion of sitter form data to variables used in API calls. This module is now utilized in the `home.tsx` and `get-sitters.ts` files. Consequently, the `SittersVariables` interface was defined in `sitter/types.ts` to standardize the expected data structure.